refactoring of collaboration logic for maintainability, scalability, and easier to read
